appmetadata-controller/pyproject.toml
[project]
name = "appmetadata-controller"
version = "1.0.0"
description = "Kubernetes Controller for ApplicationMetadata CRD"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "kopf>=1.37.2",
    "kubernetes>=29.0.0",
    "pyyaml>=6.0.1",
    "prometheus-client>=0.21.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-json-logger>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3